+++ src/pages/common/header.tsx
@@ -1,20 +1,23 @@
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router'
 
+import AccountMenu from '@/pages/common/account-menu.tsx'
+import AuthButton from '@/pages/common/auth-button.tsx'
+
 export default function Header() {
     const navigate = useNavigate()
     const [user, setUser] = useState({
         username: '',
         nickName: '',
     })
-    const [isOpen, setIsOpen] = useState(false)
+
     const fetchMy = () => {
         const stored = localStorage.getItem('userData')
         if (!stored) return
 
         try {
             const userData = JSON.parse(stored)
-            if (userData && userData.username) {
+            if (userData) {
                 setUser({
                     username: userData.username,
                     nickName: userData.nickName,
@@ -25,21 +28,13 @@ export default function Header() {
         }
     }
 
-    function viewLogin() {
-        navigate('/auth/login')
-    }
-
-    function viewSignup() {
-        navigate('/auth/signup')
-    }
-
     function viewHome() {
         navigate('/')
     }
 
     useEffect(() => {
         fetchMy()
-    })
+    }, [])
 
     return (
         <div>
@@ -49,42 +44,7 @@ export default function Header() {
                         🎬 MovieYou
                     </a>
                 </h1>
-                {!user && (
-                    <div className="space-x-3 text-sm">
-                        <button
-                            className="px-3 py-1 text-gray-700 hover:text-blue-600"
-                            onClick={viewLogin}
-                        >
-                            로그인
-                        </button>
-                        <button
-                            className="px-3 py-1 bg-blue-600 text-white rounded hover:bg-blue-700 shadow"
-                            onClick={viewSignup}
-                        >
-                            회원가입
-                        </button>
-                    </div>
-                )}
-                {user && (
-                    <div className="relative text-sm">
-                        <div
-                            onClick={() => setIsOpen(!isOpen)}
-                            className="text-gray-800 w-40 py-3 text-sm rounded-md hover:bg-gray-300 transition text-center cursor-default"
-                        >
-                            {user.nickName}님
-                        </div>
-                        {isOpen && (
-                            <div className="absolute right-0 mt-2 w-40 bg-white text-black rounded-md shadow-md border border-gray-200">
-                                <button className="block w-40 text-center px-5 py-3 text-sm hover:bg-gray-100 transition">
-                                    내 정보
-                                </button>
-                                <button className="block w-40 text-center px-5 py-3 text-sm hover:bg-gray-100 transition">
-                                    로그아웃
-                                </button>
-                            </div>
-                        )}
-                    </div>
-                )}
+                {user.nickName == '' ? <AuthButton /> : <AccountMenu user={user} />}
             </header>
         </div>
     )
